Parent AMI + clean-up + some docs
