+++ src/providers/image-builders/aws-image-builder/deprecated/index.ts
@@ -1,0 +1,5 @@
+export * from './ami';
+export * from './common';
+export * from './container';
+export * from './linux-components';
+export * from './windows-components';
