+++ src/providers/image-builders/aws-image-builder/deprecated/linux-components.ts
@@ -1,7 +1,7 @@
 import { aws_s3_assets as s3_assets } from 'aws-cdk-lib';
 import { Construct } from 'constructs';
-import { ImageBuilderComponent } from './common';
-import { Architecture, RunnerVersion } from '../common';
+import { Architecture, RunnerVersion } from '../../../common';
+import { ImageBuilderComponent } from '../index';
 
 /**
  * Components for Ubuntu Linux that can be used with AWS Image Builder based builders. These cannot be used by {@link CodeBuildImageBuilder}.
