+++ src/providers/image-builders/aws-image-builder/deprecated/windows-components.ts
@@ -1,7 +1,7 @@
 import { aws_s3_assets as s3_assets } from 'aws-cdk-lib';
 import { Construct } from 'constructs';
-import { ImageBuilderComponent } from './common';
-import { RunnerVersion } from '../common';
+import { RunnerVersion } from '../../../common';
+import { ImageBuilderComponent } from '../builder';
 
 /**
  * Components for Windows that can be used with AWS Image Builder based builders. These cannot be used by {@link CodeBuildImageBuilder}.
