+++ src/providers/image-builders/index.ts
@@ -1,8 +1,9 @@
-export * from './ami';
+export * from './api';
+export * from './aws-image-builder';
 export * from './codebuild';
+export * from './codebuild-deprecated';
 export * from './common';
-export * from './container';
-export * from './windows-components';
-export * from './linux-components';
+export * from './components';
+export * from './aws-image-builder/deprecated/windows-components';
+export * from './aws-image-builder/deprecated/linux-components';
 export * from './static';
-export * from './ng';
